+++ src/eaf/eaf.h
@@ -43,8 +43,8 @@
 #define EAF_MALLOC(WHAT, NMEMB, TYPE)                                          \
     do { WHAT = malloc (NMEMB * sizeof(TYPE));                                 \
         if (!WHAT) {                                                           \
-            Rf_error(__FILE__ ": %s = malloc (%u * %u) failed",                \
-                     #WHAT, (unsigned int) NMEMB, (unsigned int) sizeof(TYPE)); } \
+            Rf_error(__FILE__ ": %s = malloc (%u * %zu) failed",               \
+                     #WHAT, (unsigned int) NMEMB, sizeof(TYPE)); }             \
     } while (0)
 #else
 #define EAF_MALLOC(WHAT, NMEMB, TYPE)                                          \